Restructure workspace and start suspend proc-macro implementation
