+++ susync-macros/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "susync-macros"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+quote = "1.0.35"
+syn = {version = "2.0.48", features = ["full"]}
+
+[lib]
+proc-macro = true
