+++ susync/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "susync"
+version = "0.1.2"
+edition = "2021"
+license = "MIT"
+documentation = "https://docs.rs/susync/"
+description = "An util crate to complete futures through a handle. Its main purpose is to bridge async Rust and callback-based APIs."
+repository = "https://github.com/tcerqueira/susync"
+readme = "README.md"
+keywords = ["async", "callback", "bridge", "future", "handle"]
+categories = ["asynchronous"]
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+spin = { version = "0.9.8", default-features = false, features = ["spin_mutex"] }
+thiserror = "1.0.51"
+susync-macros ={ path = "../susync-macros" }
+
+[dev-dependencies]
+tokio = { version = "1.34.0", features = ["rt", "macros"] }
+tokio-test = "0.4.3"
